Internalize diagnostic helpers, reorganize tests (#2970)
Internalize four helper functions that are only used internally by
createDiagnostics(): collectValuesToRedact, scrubUrlWithValues,
createBeforeBreadcrumbHandler, createBeforeSendHandler.

Reorganize tests from two monolithic files (~1,580 lines) into six
focused files (~1,430 lines) in src/tests/plugins/core/diagnostics/:
- patterns.spec.ts: regex constant tests
- scrubUrl.spec.ts: scrubUrlWithPatterns tests
- scrubSensitiveStrings.spec.ts: string scrubbing tests
- beforeBreadcrumb.spec.ts: breadcrumb handler tests
- beforeSend.spec.ts: beforeSend handler tests
- createDiagnostics.spec.ts: integration tests

Handler tests now use createDiagnostics() with mock BrowserClient that
captures constructor options, extracting beforeSend/beforeBreadcrumb
handlers for testing. This validates the actual wiring rather than
testing factories in isolation.
